refactor: keep enterprise index owners bounded
