+++ backend-idea/src/main/kotlin/io/github/amichne/kast/idea/runtime/service/KastIdeaBackendRuntime.kt
@@ -246,12 +246,7 @@ object KastIdeaBackendRuntime {
         val server = try {
             AnalysisServer(
                 backend = backend,
-                config = ideaAnalysisServerConfig(
-                    transport = transport,
-                    limits = limits,
-                    config = config,
-                    workspaceFileCount = sourceIndexStore.loadManifest()?.size ?: 0,
-                ),
+                config = ideaAnalysisServerConfig(transport, limits, config, sourceIndexStore.loadManifest()?.size ?: 0),
                 lifecycleController = lifecycleController,
                 projectOpenController = projectOpenController,
             ).start()
