+++ cli-rs/src/semantics/symbol_query/database/mod.rs
@@ -1,3 +1,4 @@
 include!("query.rs");
+include!("lexical.rs");
 include!("graph.rs");
 include!("paths.rs");
